+++ flask_app/app.py
@@ -1,13 +1,12 @@
 from flask import Flask, redirect, url_for, render_template, request
-import spacy
 import pickle
-import pandas as pd
-import gensim
 import nltk
 from nltk.corpus import stopwords 
 from nltk.corpus import wordnet
 from nltk.stem import WordNetLemmatizer
 import string
+
+from utils import sentence_preprocessor
 
 app = Flask(__name__)
 
@@ -23,35 +22,23 @@ def team():
 def try_out():
     if request.method == 'POST':
         text = request.form.get('text')
-    
-        nlp = spacy.load('en_core_web_sm')
-        def sentence_preprocessor(text):
-            return_list = []
-            words = nlp(text)
-            for word in words:
-                word_base = word
-                word_lemma = word.lemma_
-                word_pos = word.pos_
-                word_tag = word.tag_
-                word_dep = word.dep_
-                word_length = len(word)
-                word_capitalization = str(word)[0].isupper()
-                word_punctiation = str(word).isalnum()
-                word_stop = word.is_stop
-                is_ner = str(word) in set(ent.text for ent in words.ents)
-                return_list.append((word_base, word_lemma, word_pos, word_tag, word_dep, 
-                word_length, word_capitalization, word_punctiation, word_stop, is_ner))
-            return return_list
-        df_list = sentence_preprocessor(text)
-        intermediate_df = pd.DataFrame(df_list, columns=['WordBase', 'WordLemma', 'WordPOS', 'WordTag', 
-        'WordDep', 'WordLength', 'IsCapitalized', 'NonPunctuation', 'IsStop', 'PossibleNER'])
-        pipeline = pickle.load(open('ner_model.pkl', 'rb'))
+
+        ##### NER Model #####
+
+        df_list, intermediate_df = sentence_preprocessor(text)
+
+        with open('ner_model.pkl', 'rb') as ner_f:
+            pipeline = pickle.load(ner_f)
         preds = pipeline.predict(intermediate_df)
+
         rtn_ners = []
         for pred, word in zip(preds, df_list):
             if pred == 1:
                 rtn_ners.append(word[0])
+        
+            ## add spacy ner model here. 
 
+        ##### LDA Model #####
         dct = pickle.load(open('dct.pkl', 'rb')) 
         lda_model = pickle.load(open('lda.pkl', 'rb'))
